requirements.txt
Flask
python-dotenv
openai
youtube-transcript-api
